move test mock to its own module
