+++ runtime/parachains/src/lib.rs
@@ -26,3 +26,6 @@ mod initializer;
 mod paras;
 mod scheduler;
 mod validity;
+
+#[cfg(test)]
+mod mock;
